Removed dependency on proprietary GIFStreamMetadata
When compiling and running in Java 7, a warning is shown about com.sun.imageio.plugins.gif.GIFStreamMetadata being proprietary and therefore deprecated. In a OSGi context, an external dependency was required for sun imageio, which I could not find in Maven Central. The changes in this commit should fix this.