Enable custom schema checks for linear/non-linear content
The 'properties' string passed along all checkers is used to store
temporary internal info, based on an internal vocab.

This facility is currently used to limit EDUPUB structural checks to
linear content. The non-linear nature of a content document is added
as a `epubcheck:non-linear` property to the item.

Close #539