Report links to out-of-spine items as errors
This is mandated in both EPUB 2.0.1 and EPUB 3:

- EPUB Publications 3.0.1 says (in section 3.4.12 "The spine Element"):

> All EPUB Content Documents that are linked to from EPUB Content
> Documents in the spine must themselves be listed in the spine.

- Open Packaging Format 2.0.1 says (in section 2.4 "Spine"):

> All OPS Content Documents that are part of the publication (i.e.
are listed in the manifest) which are potentially reachable by any
reference mechanism allowed in this specification must be included in
the spine.

Note: False negatives will be raised when a Content Document is in the
manifest but not linked from any other document. Checking the existence
of a chain of links from an in-spine document is difficult with the
current object model, and the case is allegedly rather rare.

See also #545.