Remove support for Java 6
From now on, EpubCheck no longer officially supports Java 6. This
commit removes Java 6 from the CI-tested environments.

Note:

This decision is due to the recent inclusion of the Twelve Monkeys [1]
ImageIO library, which is incompatible with Java 6 in its latest
version.

It is still possible to run EpubCheck on Java 6 by replacing the
TwelveMonkeys libraries with their earlier version `3.0.x`, but with
lower JPEG support on Java 8.

[1] https://github.com/haraldk/TwelveMonkeys