Fix case insensitive check of encoding declaration
An HTML meta element with `http-equiv` in character encoding declaration
state must be parsed in a case insensitive manner.
This commit fixes a Schematron typo (the `'i` parameter of the `matches`
function was unquoted, hence evaluated as an empty sequence instead
of a string, which was consequently ignored) to propertly enable a
case-insenstive comparison.

Fixes #603