Remove duplicate reporting of missing images
An image file was not declared in the OPF and not present in the
package but referenced in an XHTML document was reported twice,
as RSC-001 and RSC-007.
The duplicated report (from the extra link checker in the `ctc`
package) has been removed, in favor of the existing RSC-007.

Fixes #588