Generate XML with JVM libraries in javax.xml
This commits changes the generation of XML reports to use regular Java
libraries avoiding bad output.
It also checks for not UTF-8 characters and escapes them.
Finally, it adds the list of media-types included in the epub.

The tests have been enhanced to better compare the actual and
expected results.
Some tests cases have been added to test for encrypted or obfuscated
epubs.

Fixes #670.
Fixes #517.