fix: silence a Saxon warning (Schematron XSLT)
The Schematron XSLT built in Jing as an `/..` XPath expression which
always evaluates to the empty sequence.
Saxon 9.8 consequently raises a SXWN9000 warning, when the validators
are created (statically, during EPUBCheck's intializations).

This PR initalizes Jing's transformer factory with a static error listener
that ignores this warning.

This is tested as an integration test (since the warning is produced
on the standard error stream at initialization time), run via the
Maven failsafe plugin.

Fixes #859