test: make sure tests reset the default locale
Some tests were setting the default Locale, but were not setting it
back to the original value.
This side effect caused later localization tests to fail.

This PR fixes that by making sure the default locale is reset, in a
`finally` block, whenever it is explicitly set in tests.

Fixes #916