chore: remove `openjdk11` from the build matrix
Travis CI currently fails to install OpenJDK 11. Apparently the install
script is quite brittle and breaks whenever a new build is uploaded and
Oracle removes the old download page.

See https://travis-ci.community/t/install-of-openjdk11-is-failing/1914

Let's keep track of this issue on Travis side, and re-add `openjdk11`
as soon as it is fixed.