fix: keep on processing after failing to read GIFs
Reading some GIF images caused EPUBCheck to crash with a
`java.lang.ArrayIndexOutOfBoundsException`. See #756.

This change expands the catch block to any `RuntimeException` to at
least catch the problem and keep on processing.

Further work needs to be done to handle #756 properly, and do not
reject said GIFs as corrupted.