fix: script data blocks don't need to be declared
Content Documents with script data blocks (`script` elements with inline
data that isn't Javascript) should not need to be declared as `scripted`
in the Package Document.

`OPF_014` reports missing item properties, and already only reports a
missing `scipted` property when Javascript-typed `script` is found.

But `SCP_005` was also raised as `ERROR`, duplicating `OPF_014`, and
incorrectly _not_ depending on the script type.

This PR fixes that by suppressing message `SCP_005`.