[fix #789] use the new jhove schema
In order to fix the issue #789, the new jhove schema is implemented.

Indeed, this new schema has been published and includes an 'id' attribute as well as a 'warning' level. So the alignment between the severity attribute and the level is now straightforward.

EpubCheck level | severity attribute
----------------| ------------------
fatal error     | error
error           | error
warn            | warning
hint            | info