fix: do not require the Navigation Document to have an index in an Index Publication
This fix removes the index-specific schemas when checking the Navigation
Document.

It is probably too simple, a better implementation would only do so if
the Nav Doc is not in the spine or is not the only Content Document in
the spine, but there is no easy way to check that latter condition in
the NavChecker.
The fix is a reasonable compromise, as a false-negative (the Nav Doc
**must** have index content when it is the unique Content Document) should
be pretty rare, and is always better than a false-positive.

Fixes #1122