feat: do not report Media Overlays ordering mismatch
EPUBCheck v4.2.5 introduced a new check (MED-015) to report a mismatch
between Media Overlays and Content Document order.
But this rule will be relaxed in EPUB 3.3.

This commit preemptively demotes MED-015 to a USAGE.

See #1225. We'll decide if we want to totally remove the feature when we
fully implementat EPUB 3.3.