feat: disable usage reports related to javascript content
The `ctc` package contained checks that looked at javascript content and tried
to detect suspicious or potentially problematic usage patterns.

This check were necessarily incomplete, and based on regex pattern matching,
so inherently open to many false positives or false negatives.

This change  suppresses all of them. Inspecting javascript is not possible and
out of scope for EPUBCheck.

Note: this change will only be effective when the checks of the `ctc`
package  are disabled.