fix: improve checking of missing SVG link labels, now USAGE
This commit changes the severity of error `ACC-011` to USAGE (was WARNING).
Labelling links is not strictly a spec requirement, and checking for
missing labels should be done with specialized accessibility inspection.

The check is still available, with an improved logic: the link is
reported as having a missing accessible label only if:
- it does not have an `xlink:title` attribute
- AND it does not have an `aria-label` attribute
- AND it does not have a `title` element child
- AND it does not have a `text` element child

Note that his logic is incomplete, as it does not check that the content
of `title` or `text` children is not empty. But reporting links that
have none of the above can be somewhat useful as a simplified first-pass
accessibility check.

Fix #1353