fix: OPF-018 was incorrectly reported with inline CSS
`OPF-018` checks that the `remote-resources` property is not defined
if no remote resource was found.
Prior to this commit, it was incorrectly reported when an XHTML content
document carrying the property also contained inline CSS: the CSS checker
thought that the property was not required since no remote resource
was found in the CSS itself.
This commit fixes that bug by only checking `OPF-018` for standalone
CSS documents.

Fix #1335