feat: check that container-relative URLs have no query
EPUB relative URLs must be valid-relative-ocf-URL-with-fragment strings,
which must not have query components.

This commit adds a new check that reports `RSC-033` (error) when a query
is found in a container-relative URL.

It also slightly improves the URL checks in single-file validation mode,
to ensure a URL is not always considered remote in that mode.