feat: report as usage container resources not listed in manifest
EPUBCheck used to report any resource found in the the container but not
listed in the manifest as a warning (since #58 was fixed, in v1.1). But
the EPUB specification does not require that.

This commit downgrades the severity of `OPF-003` to a usage report.

See also #1452
See also w3c/epub-specs#563