feat: check that the manifest is not self-referencing
This commit adds a new check, reported as `OPF-099` (error), to verify
that the package document `manifest` does not include an `item` element
that refers to the package document itself.

This statement was apparently not checked by EPUBCheck previously. Worse,
the code was entering an infinite loop, as the package document checker
was creating a new checker for itself, recursively.

Fix #1453