feat: "xmp" is no longer a valid package link property
The following entries were removed from the metadata link vocabulary in
EPUB 3.3, due to low usage report:

- the "acquire" relationship
- the "xmp" property

This commit removes the "xmp" from the link property vocabulary, so any
use of it will be reported as an error.

It keeps the "acquire" property in the relationship vocabulary, for
backward compatibility notably with the EPUB Previews specification were
it is defined.

Close #1457