fix: allow unknown alternate style sheet class names
A class name found on a `link` element representing a style sheet was
reported as an error when it was not one of the known keyword defined
in [EPUB Alternate Style Tags](https://idpf.org/epub/altss-tags/).

This commit fixes that by allowing any name in the class attribute.

This commit also downgrades `CSS-005` to a usage; it is reported when
two mutually-exclusive class names are used on the same link element.
EPUB Alternate Style Tags defines these as mutual exclusive names, but
it is mostly unimplemeneted and this was never strictly forbidden.

Fix #989