fix: allow remote URLs in HTML "cite" attributes
The resources referenced in `cite` attribute URLs are not publication resources.

This commit disables the publication-resource-specific URL checks for the `cite` attribute. For instance, the referenced document is not subject to resource location restrictions.

The URL is still registered in the reference registry, so that the existence of the underlying container resource is checked (for relative URLs).

Fix #1495