fix: reject unknown `epub:` attributes in SVG
The spec only requires SVG to be well-formed, but one can argue that
attributes in the EPUB namespace can only be defined by the EPUB specs
and therefore should still be validated.

For instance, EPUB already specifies where `epub:type` is allowed in SVG,
as a normative statement.

This commit updates the persmissive SVG schema to reject all attributes
in the `http://www.idpf.org/2007/ops` namespace except the known ones.