fix: simplify `--failonwarnings` documentation
The help text presented option `--failonwarnings` as if it also allowed
values `--failonwarnings+` and `--failonwarnings-` to respectively
enable and disable warning messages.

The CLI implementation did not recognize these options. In addition, the
single option `--failonwarnings` is sufficient, since the default behavior
is to not return an error code when warnings are reported.

This commit fixes the CLI help text, and tweak the implementation to
remove traces of the legacy `--failonwarnings+` and `--failonwarnings-`
options.

Fixes #1575