fix: `mailto` URLs were reported as missing resources
Fix a regression bug introduced in PR #1582, in the implementatiom of
`OCFContainer#isRemote(URL)`.
The buggy implementation considered `mailto` URLs as local URLs. As a
result, it reported an error as no matching local resource could be
found in the container.

The new implementation says a URL is remote when it is not a `data`
URL and is not the URL of a resource in the container.

Fix #1595