feat: report unusual 'epub' namespace (usage)
Report a USAGE message when a namespace other than the usual
`http://www.idpf.org/2007/ops` URL is associated to the `epub` prefix
in XHTML content documents.

This is reported as HTM_010 (thereby repurposing this unused message ID).

This is only reported as a USAGE, since the spec does not strictly
requires that the `epub` prefix is tied to the well-known URL. But it
can help catch typos as reported in #1601.

Fixes #1601