fix: do not pollute stdout JSON with messages
The message "No errors or warnings detected." was printed to `stdout`
even when EPUBCheck was configured to print a JSON report to `stdout``,
why caused the JSON output to be invalid. This commit fixes that by
checking that JSON (or XML/XMP) is not being printed to `stdout` before
writing this message.

Fixes #1603