feat: allow `its-*` attributes (ITS) in HTML
This change proactively allows ITS attributes in XHMTL content document,
which is a new HTML extension planned in EPUB 3.4.
See w3c/epub-specs#2732

The implementation naïvely removes the attributes when pre-processing
the content sent to the validators. It does not check that the attributes
are defined and conform to the requirements in ITS 2.0.

This is a first step in resolving #1616, to prevent EPUBCheck from
totally rejecting ITS attributes.